refactor: move all BlockFetch code into Hoard.BlockFetch module
Less fragmented and more like a proper vertical module.
