+++ test/Unit/Hoard/CollectorSpec.hs
@@ -15,7 +15,6 @@ import Prelude hiding (evalState, execState)
 
 import Data.UUID qualified as UUID
 
-import Hoard.BlockFetch.Events (BlockFetchRequest (..))
 import Hoard.ChainSync.Events (HeaderReceived (..))
 import Hoard.Collector (pickBlockFetchRequest)
 import Hoard.Data.Block (Block (..))
@@ -29,6 +28,8 @@ import Hoard.Effects.Monitoring.Tracing (runTracingNoOp)
 import Hoard.Effects.Publishing (runPubWriter)
 import Hoard.Effects.Verifier (runAllValidVerifier)
 import Hoard.Types.Cardano (CardanoBlock, CardanoHeader)
+
+import Hoard.BlockFetch qualified as BlockFetch
 
 
 testBlock :: CardanoBlock
@@ -91,21 +92,21 @@ spec_Collector = do
         it "should issue request for missing block" do
             let reqs = runEff []
             reqs
-                `shouldBe` [ BlockFetchRequest
+                `shouldBe` [ BlockFetch.Request
                                 { peer = testPeer
                                 , timestamp = epoch
                                 , header = testHeader
                                 }
                            ]
   where
-    runEff :: [Block] -> [BlockFetchRequest]
+    runEff :: [Block] -> [BlockFetch.Request]
     runEff db =
         let ((), events) =
                 runPureEff
                     . runLogNoOp
                     . runTracingNoOp
                     . runWriter
-                    . runPubWriter @BlockFetchRequest
+                    . runPubWriter @BlockFetch.Request
                     . evalState db
                     . runBlockRepoState
                     . runAllValidVerifier
